# public/swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "boolean"
							}
						]
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISuccessResponse": {
				"properties": {
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "boolean"
							}
						]
					},
					"message": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IUserSignUp": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignInResponse": {
				"properties": {
					"access_token": {
						"type": "string",
						"nullable": true
					},
					"refresh_token": {
						"type": "string",
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"access_token",
					"refresh_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserSignIn": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"info": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAuthAccessTokenResp": {
				"properties": {
					"access_token": {
						"type": "string"
					}
				},
				"required": [
					"access_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuthRefreshToken": {
				"properties": {
					"refresh_token": {
						"type": "string"
					}
				},
				"required": [
					"refresh_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfile": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"first_name_EN": {
						"type": "string"
					},
					"last_name_EN": {
						"type": "string"
					},
					"first_name_TH": {
						"type": "string",
						"nullable": true
					},
					"last_name_TH": {
						"type": "string",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"date_of_birth": {
						"type": "number",
						"format": "double"
					},
					"address_EN": {
						"type": "string"
					},
					"address_TH": {
						"type": "string",
						"nullable": true
					},
					"zip_code": {
						"type": "number",
						"format": "double"
					},
					"phone": {
						"type": "string"
					},
					"image_profile": {
						"type": "string",
						"nullable": true
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"first_name_EN",
					"last_name_EN",
					"first_name_TH",
					"last_name_TH",
					"gender",
					"date_of_birth",
					"address_EN",
					"address_TH",
					"zip_code",
					"phone",
					"image_profile",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfileInsert": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"first_name_EN": {
						"type": "string"
					},
					"last_name_EN": {
						"type": "string"
					},
					"first_name_TH": {
						"type": "string",
						"nullable": true
					},
					"last_name_TH": {
						"type": "string",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"date_of_birth": {
						"type": "number",
						"format": "double"
					},
					"address_EN": {
						"type": "string"
					},
					"address_TH": {
						"type": "string",
						"nullable": true
					},
					"zip_code": {
						"type": "number",
						"format": "double"
					},
					"phone": {
						"type": "string"
					},
					"image_profile": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"user_id",
					"first_name_EN",
					"last_name_EN",
					"first_name_TH",
					"last_name_TH",
					"gender",
					"date_of_birth",
					"address_EN",
					"address_TH",
					"zip_code",
					"phone",
					"image_profile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfileUpdate": {
				"properties": {
					"id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"first_name_EN": {
						"type": "string"
					},
					"last_name_EN": {
						"type": "string"
					},
					"first_name_TH": {
						"type": "string",
						"nullable": true
					},
					"last_name_TH": {
						"type": "string",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"date_of_birth": {
						"type": "number",
						"format": "double"
					},
					"address_EN": {
						"type": "string"
					},
					"address_TH": {
						"type": "string",
						"nullable": true
					},
					"zip_code": {
						"type": "number",
						"format": "double"
					},
					"phone": {
						"type": "string"
					},
					"image_profile": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfileUpdateByUser": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"first_name_EN": {
						"type": "string"
					},
					"last_name_EN": {
						"type": "string"
					},
					"first_name_TH": {
						"type": "string",
						"nullable": true
					},
					"last_name_TH": {
						"type": "string",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"date_of_birth": {
						"type": "number",
						"format": "double"
					},
					"address_EN": {
						"type": "string"
					},
					"address_TH": {
						"type": "string",
						"nullable": true
					},
					"zip_code": {
						"type": "number",
						"format": "double"
					},
					"phone": {
						"type": "string"
					},
					"image_profile": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"user_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-authen-ts",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully signup",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserSignUp"
							}
						}
					}
				}
			}
		},
		"/v1/signin": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISignInResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"access_token": "your_access_token",
											"refresh_token": "your_refresh_token"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserSignIn"
							}
						}
					}
				}
			}
		},
...
		"/v1/auth/refresh/token": {
			"post": {
				"operationId": "AuthRefreshToken",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IAuthAccessTokenResp"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"access_token": "new_access_token"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuthRefreshToken"
							}
						}
					}
				}
			}
		},
		"/v1/auth/refresh/tokens": {
			"post": {
				"operationId": "AuthBothTokens",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISignInResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"access_token": "new_access_token",
											"refresh_token": "new_refresh_token"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuthRefreshToken"
							}
						}
					}
				}
			}
		},
		"/v1/revoke/token/{user_id}": {
			"post": {
				"operationId": "RevokeToken",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "success"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/profile/{id}": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Get Profile",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IProfile"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "5",
											"user_id": "test",
											"first_name_EN": "John",
											"last_name_EN": "Doe",
											"first_name_TH": "จอน",
											"last_name_TH": "โด",
											"gender": "male",
											"date_of_birth": 1699516723,
											"address_EN": "test address",
											"address_TH": "ทดสอบ",
											"zip_code": 27000,
											"phone": "+66939999999",
											"image_profile": "test_url"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully update",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProfile",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully delete",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/profile/{user_id}": {
			"get": {
				"operationId": "GetProfileByUserId",
				"responses": {
					"200": {
						"description": "Get Profile by user_id",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IProfile"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "5",
											"user_id": "test",
											"first_name_EN": "John",
											"last_name_EN": "Doe",
											"first_name_TH": "จอน",
											"last_name_TH": "โด",
											"gender": "male",
											"date_of_birth": 1699516723,
											"address_EN": "test address",
											"address_TH": "ทดสอบ",
											"zip_code": 27000,
											"phone": "+66939999999",
											"image_profile": "test_url"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddProfileByUserId",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully create",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileInsert"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateProfileByUserId",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully update",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdateByUser"
							}
						}
					}
				}
			}
		},
		"/v1/profile": {
			"post": {
				"operationId": "AddProfile",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully create",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileInsert"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}